Don't show context menu if there's no items
Summary:
If the creator callback to useContextMenu returns `[]`, don't show a blank context menu. Instead, return. This also skips the stopPropagation, so parent context menu handlers can be used.

In practice, this is used for right clicking on a remote bookmark to not show a menu, and instead continue to delegate to `<Commit>`'s context menu.

Reviewed By: quark-zju

Differential Revision: D55938828

fbshipit-source-id: 25e43136a76c4efeb26c89e5c818380e1d4991c1