Crash when trying to backsync changes if small repo expands submodules
Summary:
As a precaution, we don't want to allow any backsyncing (e.g. through SCS cross-repo commands) for repos that have submodule expansion enabled.

To achieve this I temporarily added a a new `large_repo_id` field to the `SubmoduleExpansionData` and updated all the callsites to make sure we can check that we're not performing backsync before running the submodule expansion code.

I started setting up unit tests for this, but it became much more complicated than expected (see diffs above) and, since this is not **needed** to start Phase #1 for a few repos (i.e. read-only, forward syncing expansion), I'll be writing tests after I finish implementing Validation of expansion (which is the purpose of this stack) and have finished updating all unit/integration tests for forward syncing.

Reviewed By: mitrandir77

Differential Revision: D55802349

fbshipit-source-id: 50ae5b9bc523aed1aeac34f03e2814c27dc044d8