website: remove ghstack page
Summary:
With dotgit mode, users should be able to use the unmodified `ghstack` just fine.
There is no need for first party support for this 3rd party tool.

`ghstack` runs low-level Git commands that will not work if we migrate the `.sl/`
Git support to not use the Git on-disk format (for the benefit of laziness or
repack-free). There are already 3 different ways to `pr submit`, `ghstack` seems
just further adding confusion.

Reviewed By: sggutier

Differential Revision: D55970086

fbshipit-source-id: 85616f7f57b8bcebec0c83e326aecd4aaee090a7