globalflags: promote --reason to global flags
Summary:
`--reason` was used by commit cloud to provide why a sync was triggered by automation.

Per team discussion, we think a way for all automations to provide easy-to-read
context would be nice, and the `--reason` flag seems a good fit. Therefore, promote
this to a global flag.

Example use-cases for ISL, the reason could be "DnD rebase", "Context menu rebase",
"quick amend", "panel amend", etc. This can be useful for other automation like
watchman or scripts too.

Choosing the CLI argv instead of an environment variable because it can be
`pgrep`-ed, tracked in metalog "commit message", and can be stacked.

Reviewed By: muirdm

Differential Revision: D55929995

fbshipit-source-id: 73862cc81c6a31ab88f15cbb7f2c0075b7753fe8