Build mononoke.scripts as noarch
Summary:
This package is part of a set of packages referenced in Chef that are trivial to dual-build. Add dual-arch builds to prepare for wider availability of aarch64 servers.

In the case of this package, mark the blob as `noarch`, since these are simple scripts with no compiled code.

For general information on the aarch64 effort, please see the Arm Enablement workplace group: https://fb.workplace.com/groups/1858357931081326

Reviewed By: andreacampi

Differential Revision: D55913634

fbshipit-source-id: acfb073cd49d627a0fd9e7de9ed9acc308aef954