fix HgQueuedBackingStoreTest getBlob
Summary:
The getBlob tests in this test don't check anything. Because `repo.hg()` doesn't add any blob into the repo and we only have two trees added to the repo. As a result the getBlob tests never get to these lines
https://www.internalfb.com/code/fbsource/[b78d569bc7c0][blame]/fbcode/eden/fs/store/hg/test/HgQueuedBackingStoreTest.cpp?lines=164%2C191

Therefore, we don't check getBlob in the two tests.
This diff adds the other files into the `repo` therefore the getBlob tests will reach to those line of the code and test the getBlob function in `HgQueuedBackingStore`

Reviewed By: genevievehelsel

Differential Revision: D56033572

fbshipit-source-id: a12d0a56bb3a56994ec21b8c1e233f51dbfcdfc2