Prevent contention on the backsync databases when backsyncing is behind
Summary:
This is a SEV follow-up for S396322:

The situation we are concerned with is a sudden influx of diffs that affect www and need backsynced at once.

Before this diff, to ensure that backsync is caught up, Mononoke attempts to backsync diffs itself.

backsyncing is a process that relies on accesses to a number of resources:
* mutable counters (to know what was last backsynced)
* bookmark update log
* globalrevs
* bookmarks table
* blobstore

When backsync is lagging, any process that attempts to backsync (in addition to the backsyncer jobs) will add load
to all these resources, creating a thundering herd effect and potentially slowing down the backsyncer which is trying
to catch-up.

Instead, only query the mutable counters to be informed once the backsync is caught up.
Use exponential backoff to minimize the load on that resource.

This change is behind [a jk](https://www.internalfb.com/intern/justknobs/?name=scm%2Fmononoke#defer_to_backsyncer_for_backsync) that is currently disabled to ensure a controlled rollout.

Reviewed By: mitrandir77

Differential Revision: D55133295

fbshipit-source-id: 2a06879cf66156006b9eef4d75ceb54eb1d59b9b