Extracting submodule expansion integration test setup into helpers
Summary:
## This stack
Make the last improvements and fix the remaining gaps in test coverage.

## This diff
Extracts the setup logic from the existing integration test into a library, so it can be reused by the SCS integrations tests I'll be adding in the next few diffs.

This will avoid a lot of copy/pasting.

You can see that the test behaviour (i.e. the output) it's pretty much the same, save for some minor formatting because of prints I added.

Reviewed By: mitrandir77

Differential Revision: D55931444

fbshipit-source-id: d97fe2486c9285edaf457d9a355aa3bd239942c7