Use buffered instead of try_buffered for blob and tree packfile stream
Summary: The packfile stream generated for tree and blob items takes the form `BoxStream<Item = Result<PackfileItem>>` so it naturally makes sense to use `try_buffered` combinator for it. However, we do not want to concurrently buffer items from the stream based on count but instead on their weight (see D55965159 for further details). This weighted buffering of futures is currently only defined for `Stream` and not for `TryStream` which is why this diff updates the code to use `buffered` instead of `try_buffered` so we can go ahead and replace it with `weighted_buffered` in a follow up diff

Differential Revision: D56015594

fbshipit-source-id: 0152ef6ab01e38f730a9dcd2684f205601073b78