Use tree_and_blob concurrency value for fetch tree_and_blob count
Summary: There was one place in the packfile stream generator code where we were hardcoding the concurrency value. This section of the code was used to determine the count of unique tree and blob objects to be included in the packfile. This diff ensures that we use the user specified concurrency value for computing it.

Differential Revision: D56015767

fbshipit-source-id: 508e2cf0c9fbb9bee488f5c98d12780ae1c75d1f