Use real ui.username as author for optimistic commit
Summary:
Previously, we inherited the optimistic commit author from the previous commit, which was typically some unrelated user when making a commit off of a public commit.

instead, let's take the username string from the sl config, and insert that first.

We can use this config value elsewhere, for example to show if a commit is made by you.

Reviewed By: quark-zju

Differential Revision: D56043322

fbshipit-source-id: d7fb00fdbfec28376ce51e6459f02e096656bdce