pypathmatcher: support downcasting intersectionmatcher
Summary: Support extracting a pure Rust matcher from a Python intersectionmatcher. You get an intersectionmatcher for the "sl files" command on an eden sparse (AKA filteredfs) repo. We didn't hit this on a non-eden sparse repo because we walk the dirstate instead of the manifest. For eden repos, however, the dirstate doesn't contain all files, so we must walk the manifest (we should probably just use manifest+status for working copy "files" in both cases).

Reviewed By: quark-zju

Differential Revision: D56031382

fbshipit-source-id: e9a91b6dc940718bf8b2db28249fa1c2b43ea4cd