FilteredFS: add code to ignore missing/invalid filter files
Summary:
# Context

Right now checking out a commit that's older than the active filter file creation commit causes the checkout operation to fail. This failure happens because Mercurial expects to be able to find the filter file in the repo, but the filter file wasn't created until a later commit.

# Solution

We should simply apply the null filter and send a warning to the user in this case. The user will be aware that a filter couldn't be applied because they're checking out an old commit, but they won't be blocked by the error.

# This diff

Adds the ability to fallback to an AlwaysMatcher if the specified filter file does not exist in the repo. A warning is logged and the next diff will introduce some ODS counters to detect when this fallback path is frequently hit.

Reviewed By: quark-zju

Differential Revision: D55781504

fbshipit-source-id: 0ec43f5e3f1f456e215be0246897ae982eced23c