FilteredFS: Add test that showcases broken checkout
Summary:
# Context

Right now checking out a commit that's older than the active filter file creation commit causes the checkout operation to fail. This failure happens because Mercurial expects to be able to find the filter file in the repo, but the filter file wasn't created until a later commit.

# Solution

We should simply apply the null filter and send a warning to the user in this case. The user will be aware that a filter couldn't be applied because they're checking out an old commit, but they won't be blocked by the error.

# This diff

This diff adds a test case that shows what happens when you try to apply a non-existent filter.

It also adds a test case that shows what happens when you check out to a commit older than the commit that added the active filter file.

Reviewed By: quark-zju

Differential Revision: D55781507

fbshipit-source-id: e519273a2878108d673adfdd0b880a36d96af56b