FilteredFS: add test for edenffi ODS counters
Summary:
# Context

Right now checking out a commit that's older than the active filter file creation commit causes the checkout operation to fail. This failure happens because Mercurial expects to be able to find the filter file in the repo, but the filter file wasn't created until a later commit.

# Solution

We should simply apply the null filter and send a warning to the user in this case. The user will be aware that a filter couldn't be applied because they're checking out an old commit, but they won't be blocked by the error.

# This diff

Adds a test for ODS counters that are added in the next diff. These ODS counters can be monitored -- if they ever spike, that would indicate that something is wrong w/ the filters we've deployed across the fleet. In other words, this ODS counter is a warning system to avoid silent filter file read failures.

Reviewed By: quark-zju

Differential Revision: D55781503

fbshipit-source-id: 43927b96aee4c569e52a2bf64349bc188502fabf