FilteredFS: clean up existing tests
Summary:
# Context

Right now checking out a commit that's older than the active filter file creation commit causes the checkout operation to fail. This failure happens because Mercurial expects to be able to find the filter file in the repo, but the filter file wasn't created until a later commit.

# Solution

We should simply apply the null filter and send a warning to the user in this case. The user will be aware that a filter couldn't be applied because they're checking out an old commit, but they won't be blocked by the error.

# This diff

I noticed some problems while writing the tests in the previous diff. This diff cleans up some of those problems.

Reviewed By: quark-zju

Differential Revision: D55781502

fbshipit-source-id: 45283e9137755504520bbe0831757b5d97307738