FilteredFS: Add ODS metrics
Summary:
# Context

Right now checking out a commit that's older than the active filter file creation commit causes the checkout operation to fail. This failure happens because Mercurial expects to be able to find the filter file in the repo, but the filter file wasn't created until a later commit.

# Solution

We should simply apply the null filter and send a warning to the user in this case. The user will be aware that a filter couldn't be applied because they're checking out an old commit, but they won't be blocked by the error.

# This diff

Adds ODS counters to the EdenFFI code. These are currently only used for FilteredFS related metrics, but we could extend these in a later diff if we end up adding more entry points to the FFI library.

Reviewed By: muirdm

Differential Revision: D55781505

fbshipit-source-id: eb420d36396294f5ccb7318a3936e1194f975c83