test: properly support test instead of eager in repo URLs
Summary:
When using `test` as an url on tests, these are intended to be used as EagerRepos, for instance `test:name` and `test://name` should be the same as `eager:$TESTTMP/server-repos/name`.

There was a bug that previously hadn't surfaced until EdenFS tests started to be used, where cloning a new repo with `hg clone` didn't work properly if `test` was used instead of `eager:TESTTMP`, which this diff fixes. The cause was the repo simply not being recognized as eager due to the URL type, and this in turn made the clone commands to behave as shallow clones.

Reviewed By: muirdm

Differential Revision: D56155730

fbshipit-source-id: bceaa8f3fd4b41c70c4d29feaf5c577896d73b0a