Mount: log BackingStore type in LogEvent
Summary:
# Summary

Now that we have multiple BackingStore types, it would be nice to know which type is being used by looking at the logs. While it's possible to figure out by looking at certain ObjectIDs and whatnot, let's make it even more clear which type is being used at all times. This will help us rule out FilteredFS in a lot of cases ;)

# Solution

We should log the type of BackingStore we're using at mount time. If possible, we should also include this information in every single EdenFSEvent and Eden CLI invocation if possible.

# This diff

Sends the BackingStore type to Scuba so that we can remotely tell which type of BackingStore a host was using. This helps in the case that we don't have EdenFS daemon logs.

Differential Revision: D56171327

fbshipit-source-id: 648f83c4f5d6c524c27d4afbb3528c99b7ddfc0f