Move LocalStoreCachingPolicy to BackingStore
Summary:
### context
`LocalStoreCachingPolicy` describs the kind of data cached in the LocalStore. As it is a BackingStore config and each type of backingstore can have a different policy, it would be less error prone if it defined inside `BackingStore`. In the prior diff in this stack we have to define it inside `ObjectStore` and `EdenServer` determined the policy based on the backingstore type.
This diff move the policy to `BackingStore` and then each backingStore based on its type set the `LocalStoreCachingPolicy`. Also, all of the BackingStores have a `getLocalStoreCachingPolicy()` method, therefore the `ObjectStore` can find the policy and use it to read the LocalStore when it is required.

Reviewed By: jdelliot, genevievehelsel

Differential Revision: D55225501

fbshipit-source-id: c59c565fd702bc0d4577e5ee8e97e7eea5607e1c