mononoke test showing problem with sync over merge&mapping change
Summary:
This test shows that even though the sync over mapping change was fixed in
D54877757 the scenario where we both merge and change mapping still needs some
work.

NOTE: this test is getting long&slow but I'm not sure if carving out a separate
test from it would make it better - most of the setup needs to be done anyway.
I'll keep that in mind when working on further diffs.

Reviewed By: andreacampi

Differential Revision: D56135497

fbshipit-source-id: 2052c536c248d56541becb250cdae6bc2fa7cace