Prefer vscode editor font for monospace
Summary:
Use the configured vscode font first for ISL instead of just a default monospace that's hardcoded. In environemtns where this css var is not defined, this does nothing.

Eventually, it'd be nice to make the vscode allow configuring some theme vars like this especially for ISL. But I think most people would want the ISL diff view font to match the normal vscode editor.

The CSS variable syntax is a little weird for fonts, apparently it only uses the first `,`, so even though the variable looks like it has a bunch of values, conceptually it's:
```
var(
  "--vscode-editor-font-family",
  "source-code-pro, SFMono-Regular, Consolas, 'Liberation Mono', Menlo, Courier, monospace"
)
```

Reviewed By: quark-zju

Differential Revision: D56162272

fbshipit-source-id: 26753eb3e2ad5504a7395e7572a74832a46e1c79