Subscribe to watchman only when ISL is opened
Summary:
Previously, when a Repository was created, we would start both dirstate (for triggering sl log) and watchman subscriptions (for triggering sl status).

This is wasteful for the vscode platform.
In vscode, we create Repositories for workspace folders. This allows the extension to provide non-ISL source control features.

However, this means we create Repositories when launching vscode, even if you don't use ISL, or have workspace folders that you don't use ISL for.

Instead, let's only make the watchman subscriptions when you start using a Repository for an ISL connection. We do this with a refcount, so that opening and closing multiple ISLs won't mess things up somehow.

Reviewed By: quark-zju

Differential Revision: D56157043

fbshipit-source-id: a586882a0cf03d7f69878d48dc156a05f458ffa9