Add staged throttler
Summary:
Add a function that can incrementally throttle an input stream.

The goal is that we can begin with no throttling, but if we see a lot of events, we can start throttling. If the burst of events stops for a while, we can go back to not throttling. Otherwise, we keep throttling while there continues to be an influx of events.

This is intended to be used as a layer between watchman fs event detections and calls to `sl status`.

This solves two problems:
- if a repo's watchman config is not set up well, there may be files which  CONSTANTLY fire continuously all the time. For example, due to a weird IDE interaction or build system.
- if there is a build process that is run occasionally, it may write lots of files to the repo really quickly. This is a burst pattern that will let up once the build is complete. We can start throttling those events to not cause unnecessary status calls, but once the build seems to be over, we can go back to not throttling.

This throttling system has a configuration of "stages" of throttling. You can choose if you want to start out with any throttling, as well as what limits are used to transition to the next stage and reset back to the initial stage.

Some minor technical notes:
- you advance to the next stage based on inputs, regardless of if they end up throttled or not. This means if you get a HUGE burst of events, you can "skip" right to the most aggressive stage of throttling, without spending much time in the in-between stages.
- resetting to the initial stage looks only at the time span since the previous emission. So a gap of NO emissions for a given time is required to reset (as opposed to no longer reaching some threshold of # of events per time span).
- You can provide on an "onEnter" callback for stages, to hook into when the throttling becomes active.

Reviewed By: quark-zju

Differential Revision: D56087417

fbshipit-source-id: bacd0d7eb4564e722a923743e57e0b9357f286b9