Provide wait-for-bookmark primitive for Mononoke Git Server
Summary: For EdenAPI and SCS, we have utility functions in `library.sh` that allow us to wait until the latest bookmark updates have been captured. For EdenAPI, we use `hg` and for SCS we use `scsc` and neither of these are available for `Mononoke Git Server`. This diff adds a `wait-for-bookmark` utility function relying on the `git ls-remote` command. This will be used in the immediate next diff.

Reviewed By: gustavoavena

Differential Revision: D56187330

fbshipit-source-id: daa16097727c4f3d469e810fc44877638b47f178