Use NoopBookmarksCache in GitExport
Summary:
`GitExport` starts off with a `MononokeApp` constructed repo for the input repo (like`fbsource`, `configerator`, etc.) and then goes on to manually construct a temporary repo for exporting part of the source repo as a Git repo. This manual construction involves using the `TestRepoFactory` that by default uses `WarmBookmarksCache` as the implementation of `BookmarksCache` facet for the repo. `GitExport` should not use `WarmBookmarksCache` since it won't ever update the created bookmarks and uses them for one-time repo export.

This diff adds support in `TestRepoFactory` to specify a custom implementation of `BookmarksCache` to be used while creating a repo. `GitExport` relies on this to use `NoopBookmarksCache` instead of `WarmBookmarksCache`. This solves integration test failures when I start using `BookmarksCache` in `Mononoke GRit Server`

Reviewed By: gustavoavena

Differential Revision: D56191662

fbshipit-source-id: 8b527008979ae0e29c0fb10260e371058ec33d51