scmstore: remove aux local store
Summary: The "local" stores are inside .hg and are durable, whereas the "shared" stores are in hgcache and can be deleted. The local stores are mainly used when creating commits locally (the data needs to be stored durably since the server doesn't have the commit information). However, aux data is always optional, so it doesn't really make sense to have a local store. When reading aux data, we basically never get a hit in the local store since aux data is almost always written to the shared cache. Removing the local aux store saves some work and bookkeeping during fetches.

Reviewed By: zzl0

Differential Revision: D56141723

fbshipit-source-id: a1af8fded816ab6baa04c2b7a8b06d7eed422f11