tests: make cd error out in certain cases
Summary:
Normally when running `cd` on the old .t tests and also when running cd on a
terminal this command errors out if a directory does not exist. Before this
diff trying to cd in a non existing directory wouldn't error out, which
sometimes could lead to some unexpected behaviors.

Reviewed By: muirdm

Differential Revision: D56193533

fbshipit-source-id: bbee3f37c5b85166628768d5e0edf2aa771ee142