privhelper: set the fstype to fuse.edenfs
Summary:
We are teaching eden to expose itself as fuse.edenfs so that fuse.edenfs can
be added to the PRUNEFS option of updatedb to prevent it from crawling an
EdenFS repo.

There were a few references to the vfstype in code, which we have fixed. But
just in case there is some other reference we can't find we want to rollout this
change with a config option to make it easy to put back.

The privhelper does not read the eden config directly, so we need to pass the
config value over the socket. Sometimes tests talk to the privhelper so its
important that the new eden daemon can still talk to the old privhelper.
Unfortunatly, the old privhelper is gonna error if we try to pass an extra value
it's not expecting. We could rollout the changes to the privhelper and then
wait to teach eden to pass the extra vfs type. But that will be slow. Instead
we can retry failed requests without the extra arg. This will allow the new
eden to still work with the old privhelper.

Generally the privhelper and eden process are both updated at the same time
in production, so the request should succeed on the first attempt.

Reviewed By: jdelliot

Differential Revision: D55531178

fbshipit-source-id: 8e085e26b8ee94ae657a22f93e9fb0ae3fdc5a24