Provide InMemory large repo for submodule expansion
Summary:
## This stack
See summary of D55701183.

## This diff

See D55701183 for more context.
This diff starts using the `InMemoryRepo` to pass the large repo to the submodule expansion and validation primitives.
In the next diff (D56137862) the large repo will be used to ensure that the expansion is entirely deleted when the submodule metadata file is deleted.

Reviewed By: RajivTS

Differential Revision: D56138496

fbshipit-source-id: f7f284e37ec4b6b79890c426a1d393aa26489456