Introduce InMemoryRepo container
Summary:
## This stack
Continues the implementation of the primitives that will validate submodule expansion in large repo's. bonsais.

The goal of validation is to make sure that **the submodule metadata file is always kept in sync with the expansion**.
How this will be done is more complicated than it seems, but on a high level we will
1 Get the fsnode in the submodule repo from the commit in the metadata file
2. Derive the fsnode for the new commit
3. Get the entry for the expansion and compare with the fsnode from step #1.
4. We'll perform many validations of consistency in this process (e.g. if expansion is changed, metadata file has to be changed as well).

## Challenge of validation
To efficiently verify that the working copies match, we want to get fsnodes of both the commit being expanded in the submodule (easy) and the new commit being synced to the large repo.
The second one is a problem, because in order to get it, we would need access to the large repo's blobstore and to derive the fsnode using the large repo's derived data.

The problem with is that the primitives that rewrite commits (thus expand the submodules) are, on purpose, only given the target repo's id, instead of the actual repo (and its blobstore). This was done a while back (D43002436) to make sure we don't accidentally write the target repo.

We want to keep the maintain that safety net, so the problem this diff tries to solve is:
**provide READ-ONLY access to some parts of the large repo to the commit sync primitives** (e.g. CommitInMemorySyncer).

## This diff
Introduces a new Repo facet container, called `InMemoryRepo`, which can be passed to the validation primitives and will ensure that any writes performed during validation only happen in memory, i.e. are not persisted to the large repo's blobstore or tables.

This InMemoryRepo needs to have `RepoDerivedData` and `Changesets`.
But to create a `RepoDerivedData`, I need to provide a bunch of other things, most of which I won't need.
So instead of implementing a "wrapper" for all of these things (e..g `BonsaiGitMapping`, `BonsaiHgMapping`), I defined a dummy data type that will implement these to satisfy the type checker.

This is good because it will be very clear what the validation primitives will be using and if in the future this changes, the tests will crash because they'll try to read from or write to storage that they're not supposed to. In which case the person making the change will implement a wrapper to ensure the operations only write to memory.

Reviewed By: markbt

Differential Revision: D55701183

fbshipit-source-id: 706b8d4cbef8cde6a9562a52009db90901d4a685