Generate metadata file for recursive submodules
Summary:
## This stack
See summary of D55701183.

## This diff
Generate a x-repo submodule metadata file for recursive submodules as well.

Generating only for top-level submodules was a mistake, because we don't have the necessary information to **validate recursive submodules**.

This just moves the call to `generate_additional_file_changes` to `expand_git_submodule`, which is called for every submodule (top-level and recursive).

Reviewed By: andreacampi

Differential Revision: D55966271

fbshipit-source-id: 539d799d1d7c10a8139e30caf68dee39dab553f2