Submodule expansion validation: compare expansion and submodule fsnodes
Summary:
## This stack
See summary of D55701183.

## This diff
Implement steps 2 and 3 of validation, which involves getting the fsnode of the submodule commit (from the submodule repo) and the fsnode of the submodule expansion in the large repo and comparing them.

If they are an exact match, the expansion is valid. If they're not, the expansion could contain recursive submodules. The logic to handle recursive submodules is big, so I'm adding it in the next diff.

Reviewed By: RajivTS

Differential Revision: D56179689

fbshipit-source-id: 9ed2db4525df9f2477566c28bbe4e8f2ad071b72