scmstore: simplify tracking of lfs pointers
Summary:
LFS files are fetched in two stages: first you fetch the "pointer" which contains metadata including the URL from which to download the full file. Sometimes you can't fetch the pointer and data at once. For example, the pointer exists in a local cache, but the file contents have aged out of the cache. In this case you need to fetch the file contents from the remote LFS store.

Previously scmstore would also track the pointer "origin", which was whether the LFS pointer was found in the "local" (i.e durable repo) store, or "shared" (i.e. deletable cache) store. This info was used when downloading the LFS contents. If the pointer was from the shared store, we save the contents to the shared store, and vice versa. However, I don't think it is possible to have a dangling "local" pointer. Data in the "local" store is durable and never deleted, so if the local store has an LFS pointer referenced by some commit but no LFS blob, that should be an error. So, remove the origin tracking, and always download LFS content to the shared store.

I also simplifed the cleanup of the tracked lfs_pointers to only happen after downloading remote contents (previously we would try to clean up pointers whenever we found any file contents, but I don't think that is necessary).

I initially started looking at this due to borrow checker issues trying to avoid copies as we fetch data. Getting rid of the pointer_origin mutations avoids some of the issues.

Reviewed By: quark-zju

Differential Revision: D56141713

fbshipit-source-id: 88071b0ee26af8340ba3adfae8da225ea7cd1776