scmstore: chunk up humungous prefetch calls
Summary:
It is hard to avoid allocating lots of memory when you are fetching 10 million files in one call. Let's chunk down to 100k files for smaller allocations as we fetching stuff. I added a config knob scmstore.max-prefetch-size to control this value.

Perf wise (including this entire stack), a "no-op" prefetch where all the data is available is down from 267s to 71s. Peak "anon" memory usage is down from 27GB to 8GB. 8GB is still stupidly large, but I think we need to rewrite prefetch to get that down further.

Reviewed By: quark-zju

Differential Revision: D56141718

fbshipit-source-id: b633117232cd5aa2984e8be3fb645a6c280c1190