Add more unit tests for submodule expansion
Summary:
Adding a few unit tests after syncing with mitrandir77.

One unit test just ensures that we can manually update the submodule metadata file if we make sure that the working copy matches the commit we're pointing to.

The other tests are for **very unlikely scenarios**, but that could happen in the future.
See comments in the test.
We basically want to make sure we fail submodule expansion, instead of validation, because it would mean that the commit is never synced to the large repo.

I started working on a fix in D56242513, but it didn't work straight away and, since I don't think this is a blocker for the merge, I created T186047910 as a follow up.

Reviewed By: mitrandir77

Differential Revision: D56240843

fbshipit-source-id: 4f699f59226316b9cdb1efcd0ebbda0729eddf32