Recursively validate submodule expansion fsnodes
Summary:
## This stack
See summary of D55701183.

## This diff
Properly validate the working copy of submodule expansions that have recursive submodules.

The comments for each step in `validate_working_copy_of_expansion_with_recursive_submodules` should explain the logic behind the code.
I'm basically expecting a specific structure for each possible case when processing an entry in the expansions manifest and either making the appropriate recursive call or failing validation.

LMK if the comments are not enough to explain everything and I'm happy to provide more details.

Reviewed By: mitrandir77

Differential Revision: D55969227

fbshipit-source-id: efe253aad78725e68a8154bab02e688b8a9b2b25