Fix amending when clicking submit on an individual commit
Summary: The submit button is reused for amend & submit, commit & submit, and "just" submit. If you "just" submit by clicking submit on a commit you're not checked out on, it would run through the same code path. This was checking if you had anything to amend, and amending if so. That's clearly wrong. We should only amend if you're clicking "amend & submit". We just need to check that the commit is the dot commit.

Reviewed By: muirdm

Differential Revision: D56325798

fbshipit-source-id: 6b03516b8cb826c5be5383c1c4e1da3affc2e77e