Use memory bound for processing tree/blob packfile stream
Summary:
In order to drastically reduce the possibility of OOMs, we want to restrict the concurrency of the futures in the packfile stream through not just the future weights but also through the RSS size of the process. Through observation in production we have noticed that when the RSS size of the host hits 40 GB, the task OOMs. This diff sets an upper bound of 36 GB beyond which we do not schedule futures in bulk to avoid overwhelming our memory but we also do not stop completely. This helps us make progress without blowing up.

This would result in slower clones for the users. If that does end up happening, then its an indication that we don't have a good capacity/load balance in our Mononoke GRit tier.

Note that I will make this threshold configurable in follow-up diffs

Reviewed By: andreacampi

Differential Revision: D56240689

fbshipit-source-id: e2b1ba0ab80655c107be267262df42412993c1ab