Use configurable memory bound for generating packfile stream
Summary: D56240689 updated the packfile stream generation code to add the constraint of RSS memory to be honored while processing the stream. However the constraint was a hard coded value and would require redeployment if we ever wanted to change it. This diff instead makes use of configuration to define the bound value which can be easily changed at runtime

Reviewed By: andreacampi

Differential Revision: D56244588

fbshipit-source-id: a65bd3c8281e8313c02f80bcf4e05506d7570fc2