remove Option types from Wire types for AUX data
Summary:
remove Option types from Wire types for AUX data

The proc macro was used to allow for the API types to get rid of the Option. At
first, we still generated the Wire types identical as before. However, after
testing, I have realised in binary format, those types are actually represented
in the same way if the field is always present (our case), so it is therefore safe to remove.

See the test plan. Since this diff is an actual change on the Wire side, tested all scenarios manually, since CI wouldn't cover it.

        #[serde(rename = "4", default, skip_serializing_if = "crate::wire::is_default")]
        pub blake3: <Blake3 as crate::ToWire>::Wire,

        #[serde(rename = "4", default, skip_serializing_if = "crate::wire::is_default")]
        pub blake3: <Option<Blake3> as crate::ToWire>::Wire,

Reviewed By: markbt

Differential Revision: D56350818

fbshipit-source-id: 8c920fcf917155c841e4b663eb85a9c646823b6c