revert: always run pathcopies for revert operation
Summary:
Below if condition was introduced in D31814398, at the time copy tracing was O(repo) time complexity. And we only disabled copy tracing for 'revert', and still enabled it for 'backout'.

```
if forcecopytracing or repo.ui.config("experimental", "copytrace") != "off"
```

I think it should be fine to enable it for 'revert' as well now for below reasons:
1. copy tracing is not O(repo) now, we implemented Rust version of `path_copies()` and the time complexity is controlled by come configs. See details in D55857265.
2. It is rare that `revert` will revert a long range of commits.
3. Another advantage is that revert now will not lose the copy information.

Reviewed By: quark-zju

Differential Revision: D55317446

fbshipit-source-id: 2175aa12d55fe4477c5387db5a4e21cde4d14ede