copytrace: handle 'cp' case in _filtercopies()
Summary:
`_filtercopies()` checks both dest and src sides, which can wrongly filter out 'hg cp' cases. For example, if src side did 'hg cp a b', then `a` didn't change and still exit in `base` and `src`. In this diff, if one side rename/copy a file, we only check the base file in the other side.

This changes `test-fb-ext-copytrace-amend.t` to the version before D14195031, which seems more desirable.

Reviewed By: quark-zju

Differential Revision: D56313925

fbshipit-source-id: 3fec05940de77e8afe77631ee1926b4c0149988a