Allow fetching manual stables
Summary:
Now that we have a typeahead for stables, actually save entered stables from the typeahead into persisted storage, and send them whenever fetchinging the latest stables.

Then the server can fetch the latest value for all these stables, and include those commits in the next log call.

We should be resilient to stables not being fetchable. So any given stable may succeed / fail, but the rest are not affected.

If there is an error, you can see it in the UI.

You can also delete entered stables from the fetch.

Reviewed By: quark-zju

Differential Revision: D56334759

fbshipit-source-id: 54b6a1158379663b94e3640d1af7594fa936227e