Create a ConfigParser utility to ensure validity of Mononoke Configs
Summary: A while back we had a SEV 2, where the configerator config change landed before the code config change. Since this is essentially a timing problem that can only be resolved by manually checking if the pre-requisites have completed in order, it would help to have some automation in place that can help detect such mismatch. We have already started using smart canaries with config hot reloading but that doesn't work for tools/jobs that use a static version of the config. For all such cases, this tool will help

Reviewed By: mitrandir77

Differential Revision: D56419473

fbshipit-source-id: 58e72e6d34b0b6662a08c989f0455a82b8a2162b