don't backsync merge commits not applicable to small repos
Summary:
The large repo merges synced as empty commits from large to all the small repos cause the issue where the commit equivalent to large repo master is not the small repo master but empty commit in front of it.

This issue was flying under the radar but extra validation for handling the version in forward syncing (D54965081) catches that and makes the forward syncer refuse to move forward on such changes.

Reviewed By: gustavoavena

Differential Revision: D56140281

fbshipit-source-id: 0090ac8cd6b8752dbf63f79acff364d29aefe785