sunset sha256 hash from aux data
Summary:
sunset sha256 hash from aux data

We should like to deprecate sha256 hashes from AUX data in effort to cut number
of different hashing schemes we support. This will also allow use to make the
Augmented Manifest smaller.

The only current usage of this AUX data is under a config that is not enabled.

That code path, if we need it, could migrate to blake3 hashes.

As we have confirmed while deprecating content_id from the AUX data, removing
fields from the EdenApi types using "auto wire" is safe.

Reviewed By: quark-zju

Differential Revision: D56389024

fbshipit-source-id: f42c3c7c504d658e5963a051e705bb84c920e1f4