bonsai_globalrev_mapping: use RendezVous when fetching globalrevs
Summary: Uses RendezVous when fetching globalrevs by bonsais in order to batch sql queries to decrease the load on mysql. This is necessary because in some servers like SCS we fetch globalrevs individually for every commit returned to clients when globalrevs are requested, which results in a high number of concurrent mysql queries.

Reviewed By: lmvasquezg

Differential Revision: D56476229

fbshipit-source-id: c698c825166e0b3ba2c615de6bf9a415c286cb68