Open merge tool when clicking an unresolved file
Summary:
If you click the filename of an unresolved file, open the configured merge tool (if any), instead of just opening the file normally.

This way you can choose which files to use the configured merge tool for, instead of resolving all of them one by one when clicking the button added in the last diff.

Note: we could instead have a separate button to open in the merge tool, instead of just overriding the main filename click.

Reviewed By: quark-zju

Differential Revision: D56442762

fbshipit-source-id: b9ddbc790a8402ceea312fec21aacd2eb2b58a4c